terminus-font: Fix ucstoany.awk POSIX compatibility
POSIX says that numeric strings are those that can be parsed with strtod, or
those that would be recognized as a NUMBER token, depending on the
implementation. strtod supports hex strings, but POSIX says that implementations
are not required to allow hex literals as NUMBER tokens.

The awk used by oasis (onetrueawk) does appear to support hex NUMBER tokens, but
gawk does not without --non-decimal-data or --posix.