alsa-lib: Remove snd_pcm_hwsync warning
alsa-lib calls this internally, producing a warning when linking any
binary with libasound.a. On a normal system, libasound.a is linked in
with -lasound, which seems to silence the warning. However, it is much
easier for us to just pass the whole path so instead remove the
offending warning section with objcopy.